clean up retry logics
